Fix bug when passing a relative path to to UnregisteredClassLoader
The `UnregisteredClassLoader` uses `LoadLibrary` under the hood, which has
known issues with passing in relative paths. Fix that issue by getting the
absolute path before passing to `LoadLibrary`.

Adding a unit test caught an additional bug where the loader wasn't
holding on to the `SafeHandle`, which could cause the library to stay
loaded and / or get unloaded from underneath the user. The fix is to make
the loader hold on to the `SafeHandle`, and implement `IDisposable`.